TECH-286: updated shared libs
