+++ package.json
@@ -31,13 +31,13 @@
         "@apollo/react-hooks": "^3.1.5",
         "@babel/polyfill": "^7.2.5",
         "@babel/runtime": "^7.12.5",
-        "@jahia/data-helper": "^0.0.26",
+        "@jahia/data-helper": "^1.0.0",
         "@jahia/design-system-kit": "^1.1.8",
         "@jahia/icons": "^1.1.1",
         "@jahia/moonstone": "^0.17.4",
         "@jahia/moonstone-alpha": "^0.1.3",
         "@jahia/react-material": "^3.0.1",
-        "@jahia/ui-extender": "^0.5.4",
+        "@jahia/ui-extender": "^1.0.0",
         "apollo-cache-inmemory": "^1.6.5",
         "apollo-client": "^2.6.8",
         "apollo-link": "^1.2.13",
